add: annotation in custom hooks, code: refactoring
